+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/menu-selection/index.tsx
